backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.7
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
